STABLE: Synchronization of trunk and beta branch, revisions [5938] to [6651].

+++ L2J_Server/java/com/l2jserver/gameserver/model/actor/stat/ControllableAirShipStat.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2013 L2J Server
+ * Copyright (C) 2004-2014 L2J Server
  * 
  * This file is part of L2J Server.
  * 
@@ -34,7 +34,7 @@ public class ControllableAirShipStat extends VehicleStat
 	}
 	
 	@Override
-	public float getMoveSpeed()
+	public double getMoveSpeed()
 	{
 		if (getActiveChar().isInDock() || (getActiveChar().getFuel() > 0))
 		{
